chore(example): rename app to hoppix with bundle id com.bamlab.hoppix
