+++ package.json
@@ -49,7 +49,7 @@
     "@typescript-eslint/typescript-estree": "5.61.0"
   },
   "scripts": {
-    "start:example": "yarn workspace RabbitApp start",
+    "start:example": "yarn workspace hoppix start",
     "test:core": "yarn workspace react-tv-space-navigation jest",
     "test:types": "yarn workspaces foreach run test:types",
     "test:lint": "eslint . --ext .js,.jsx,.ts,.tsx --report-unused-disable-directives --max-warnings 0 --cache --cache-strategy content --cache-location .cache/eslint.json",
