+++ packages/example/android/app/src/main/java/com/bamlab/hoppix/MainActivity.java
@@ -1,4 +1,4 @@
-package com.rabbitapp;
+package com.bamlab.hoppix;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -70,7 +70,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "RabbitApp";
+    return "Hoppix";
   }
 
   /**
