+++ packages/example/android/app/src/main/java/com/bamlab/hoppix/MainApplication.java
@@ -1,4 +1,4 @@
-package com.rabbitapp;
+package com.bamlab.hoppix;
 
 import android.app.Application;
 import com.facebook.react.PackageList;
