+++ packages/example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "RabbitApp",
-  "displayName": "RabbitApp"
+  "name": "Hoppix",
+  "displayName": "Hoppix"
 }
