+++ packages/example/ios/Hoppix.xcodeproj/xcshareddata/xcschemes/Hoppix.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1130"
+   LastUpgradeVersion = "1210"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "RabbitApp-tvOS.app"
-               BlueprintName = "RabbitApp-tvOS"
-               ReferencedContainer = "container:RabbitApp.xcodeproj">
+               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+               BuildableName = "Hoppix.app"
+               BlueprintName = "Hoppix"
+               ReferencedContainer = "container:Hoppix.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -32,10 +32,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "RabbitApp-tvOSTests.xctest"
-               BlueprintName = "RabbitApp-tvOSTests"
-               ReferencedContainer = "container:RabbitApp.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "HoppixTests.xctest"
+               BlueprintName = "HoppixTests"
+               ReferencedContainer = "container:Hoppix.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -54,10 +54,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "RabbitApp-tvOS.app"
-            BlueprintName = "RabbitApp-tvOS"
-            ReferencedContainer = "container:RabbitApp.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "Hoppix.app"
+            BlueprintName = "Hoppix"
+            ReferencedContainer = "container:Hoppix.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -71,10 +71,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "RabbitApp-tvOS.app"
-            BlueprintName = "RabbitApp-tvOS"
-            ReferencedContainer = "container:RabbitApp.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "Hoppix.app"
+            BlueprintName = "Hoppix"
+            ReferencedContainer = "container:Hoppix.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
