+++ packages/example/ios/Hoppix.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
