+++ packages/example/ios/Hoppix/AppDelegate.mm
@@ -43,7 +43,7 @@ RNKeyEvent *keyEvent = nil;
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-  self.moduleName = @"RabbitApp";
+  self.moduleName = @"Hoppix";
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
