+++ packages/example/ios/Hoppix/Images.xcassets/AppIcon.appiconset/Contents.json
