+++ packages/example/ios/Hoppix/Images.xcassets/Contents.json
