+++ packages/example/ios/HoppixTests/HoppixTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface RabbitAppTests : XCTestCase
+@interface HoppixTests : XCTestCase
 
 @end
 
-@implementation RabbitAppTests
+@implementation HoppixTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
