+++ packages/example/ios/HoppixTests/Info.plist
