+++ packages/example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "RabbitApp",
+  "name": "hoppix",
   "version": "0.0.1",
   "private": true,
   "scripts": {
